🔧 Fix dependency versions and ethers.js imports for Vercel deployment
## Dependency Updates
- ✅ All dependencies are up-to-date and compatible
- ✅ Stripe packages updated to latest versions
- ✅ Ethers.js v6 properly configured

## Ethers.js v6 Compatibility
- ✅ All ethers.utils.* calls updated to ethers.*
- ✅ Web3Provider updated to BrowserProvider
- ✅ Async signer initialization implemented
- ✅ Buffer polyfill properly configured

## Code Quality Improvements
- ✅ Fixed react-refresh issue by separating useWeb3 hook
- ✅ Updated all import paths for useWeb3 hook
- ✅ Fixed unused parameter warnings
- ✅ Build process optimized and verified

## Vercel Deployment Ready
- ✅ All imports compatible with Vercel build process
- ✅ No environment variable dependencies
- ✅ Static assets properly configured
- ✅ Build process tested and working
