+++ RentChain-frontend/src/components/common/RoleGuard.jsx
@@ -1,5 +1,5 @@
 import React, { useEffect } from 'react';
-import { useWeb3 } from '../../context/Web3Context';
+import { useWeb3 } from '../../hooks/useWeb3';
 import { useNavigate } from 'react-router-dom';
 
 export default function RoleGuard({ requiredRole, children }) {
