+++ RentChain-frontend/src/pages/Dashboard.jsx
@@ -1,6 +1,6 @@
 import React, { useEffect } from 'react';
 import { useNavigate } from 'react-router-dom';
-import { useWeb3 } from '../context/Web3Context';
+import { useWeb3 } from '../hooks/useWeb3';
 import { useTranslation } from 'react-i18next';
 
 export default function Dashboard() {
